# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "financialPortfolioApiBackend",
        "description": "API Documentation for your application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify user's email address",
                "description": "Verifies the user's email address using the provided ID and hash from the verification link. This is typically a web route.",
                "operationId": "2295166070a7e08a04399e54af513a63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The user's ID.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "The verification hash.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to /home on successful email verification."
                    },
                    "401": {
                        "description": "Unauthorized - if the user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - if the link is invalid or expired (e.g., 'signed' middleware failure).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This email verification link is invalid or has expired."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/email/resend": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Resend email verification link",
                "description": "Sends a new email verification link to the authenticated user's email address.",
                "operationId": "b91a622d7d4db17dd3d2cd9261101016",
                "responses": {
                    "200": {
                        "description": "Verification email resent successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "A fresh verification link has been sent to your email address."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - if the user is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - if the user tries to resend too frequently (throttled).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Too Many Attempts."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AuthenticationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Falha na autenticação."
                    },
                    "error": {
                        "type": "string",
                        "example": "Credenciais inválidas."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}